refactor: coupon 할인 적용 로직 훅 제거하고 유틸 함수로 분리
